requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Supabase client
supabase==2.0.2
postgrest==0.13.0

# Validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1

# Production server
gunicorn==21.2.0

# Google OAuth
google-auth>=2.22.0